requirements.txt
streamlit==1.32.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
pandas==2.2.0
python-dotenv==1.0.0
jsonschema==4.21.1
extruct==0.16.0
validators==0.22.0